Hybrid captcha mode.
When a worker encounters a captcha it will attempt a manual solve several times before checking if 2captcha is available.
--manual-captcha-solving-max-attempts sets the number of attempts each account will try to solve captcha.
If --manual-captcha-solving-max-attempts or -mcsma is set to 0 it will disable manual captcha processing. Default is 3 attempts.

Changed the captcha notification message that is sent to webhooks to include status_name, attempts (current number of attempts) and time (time remaining to solve captcha - only sent if status = 'encounter').