Bug fixes.
Manual captcha solving bookmarklet page now uses the mcs-domain instead of javascript's 'window.location'.

search.py: if an exception is caused before getting an item from search_items_queue, the code would throw an uncaught exception - 'step' not defined.
Now it checks if a search_item was fetched or not to avoid throwing an exception.